travis: tweak build matrix to separate tests and deploys
with the precious setup, we'd get some spurious actions both on PRs
and tagging.

on a PR, unit tests would run (good) but we would also get 2 unnecessary
builds for 'deploy'.

OTOH at tagging - unit tests would run again.

this is not a blocker but suboptimal, so try to split off to another
matrix var/job - 'deploy', and try to switch on job type via an if (next commit -
this is just to test out the matrix).

changelog: none

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>