travis: build stages + backend integration tests
actually, this is mostly about build stages.

the refactor is not pretty because build stages vs env matrix seems
like a half-done feature. most notably - defining a global matrix
with os/arch combinations doesn't help - jobs inherit only the first
entry in the matrix. it's also not possible to define matrices inside jobs.

the only way is to copy and paste a job N times, with different env vars.
for most jobs this is ok - esp. with a helper bash script. unfortunately
with deploy stage, one has to repeat large blocks of yml, as it's impossible
to push provider defs into a helper.

there is a placeholder for backend testing jobs but the cli is not
yet a part of the integration setup (no dockerfile, compose, etc.).
first decide how we want to include/test it, then actually add a
section here.

changelog: none

Signed-off-by: mchalczynski <marcin.chalczynski@rndity.com>