Add(Makefile): Install target for the auto-completion scripts
This creates two targets in the Makefile.

One which builds the auto-completion scripts and adds them to the local
./autocompletion/ directory.

And second an install target for the auto-completion scripts, which relies on
the build step, and installs the Bash completion scripts by default, and the
zsh-autocompletion scripts if zsh is present.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>