Add missing changelogs
The PRs:
* https://github.com/mendersoftware/mender-cli/pull/79
* https://github.com/mendersoftware/mender-cli/pull/78
* https://github.com/mendersoftware/mender-cli/pull/77
* https://github.com/mendersoftware/mender-cli/pull/76

Are all missing Changelogs. This empty commit amends this.

Changelog: Fix login with password, and improve the configuration file handling
Changelog: Add 'artifact list' command to list Artifacts on the Mender server
Changelog: Prompt for the users username if not provided on the CLI
Changelog: Add '--version' option to display the current mender-cli version

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>