fix: automatically handle reconnections in port-forward
In case of errors, instead of stopping the process and exiting,
automatically reconnect to the device and continue port-forwarding the
connections. This improves the user experience in case of temporary
errors when port-forwarding an HTTP server, where each request is a new
connection. Additionally, ignore errors on connection close.

Changelog: Title
Ticket: MEN-5565

Signed-off-by: Fabio Tranchitella <fabio.tranchitella@northern.tech>