ci: Orchestrate the release through the new GitLab CI/CD compoments
Replacing and modifying the partially implemented process.

See:
* https://github.com/mendersoftware/mendertesting/blob/c21eb2185074ec381808de2f4f693411f4c41820/templates/README.md#example-complete-release-setup

One notable change from the original tentative implementation is the
removal of `v` style versions. While this is desirable we don't have
full support for it in other parts of Mender ecosystem. See related
tasks in iteration 2 of the CCC release at QA-1161.

Ticket: QA-987

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>